Test against Redis breaking change on .NET 9
