+++ src/Security/test/DataProtection.Redis.Test/Steeltoe.Security.DataProtection.Redis.Test.csproj
@@ -6,6 +6,25 @@
   <Import Project="..\..\..\..\shared.props" />
 
   <ItemGroup>
+    <!--
+      This reference is repeated to use different versions per target framework.
+      In .NET 9, LUA scripting is no longer used, which is a breaking change that requires different mocks.
+      See https://github.com/dotnet/aspnetcore/pull/54689.
+    -->
+    <PackageReference Include="Microsoft.Extensions.Caching.StackExchangeRedis" Version="$(MatchTargetFrameworkVersion)" />
+  </ItemGroup>
+
+  <ItemGroup>
     <ProjectReference Include="..\..\src\DataProtection.Redis\Steeltoe.Security.DataProtection.Redis.csproj" />
+  </ItemGroup>
+
+  <ItemGroup Condition=" '$(TargetFramework)' == 'net8.0' ">
+    <Compile Remove="**/*.net90.cs" />
+    <None Include="**/*.net90.cs" />
+
+  </ItemGroup>
+  <ItemGroup Condition=" '$(TargetFramework)' != 'net8.0' ">
+    <Compile Remove="**/*.net80.cs" />
+    <None Include="**/*.net80.cs" />
   </ItemGroup>
 </Project>
